PP-2628 Add simple Jenkinsfile, no end-to-end tests.
At this stage, this is a bare-bones repo. In order to integrate with
the end to end tests, we need to build, unit test and push a container
to Docker Hub.

This Jenkinsfile should hopefully be just enough to get that done.

As the code evolves and we have something to integrate with the
end-to-end tests, then this Jenkinsfile can have the additional steps
added.