PP-2672 `PaymentCreator` service taking care of all concerns regarding payment creation
`PaymentCreator` service takes care of all concerns regarding the creation
of a new payment for a given product. The sole entry point of the service
`doCreate` expects the `id` of a product for which a new payment will
be created and returned.

1. The initial pre-operation step looks up the product and adds to it an
empty payment with the status of `CREATED`. If the product doesn't exist a
`PaymentCreatorNotFoundException` is thrown.

2. The actual operation then submits a payment creation request to
PublicAPI which maps the outcome to our `PaymentEntity`. The
status is updated to either `SUCCESS` or `ERROR` depending on whether
the request is successful or not.

3. Finally the post-operation saves the payment updates. If the payment
operation fails an `PaymentCreatorDownstreamException` is thrown.