PP-2672 New endpoint exposing the functionality of creating a payment
The new endpoint exposes the functionality of creating a payment for a
given product by submitting a POST request to URL:

```POST /v1/api/products/{productId}/payments```

The endpoint delegates almost the entire operation to `PaymentCreator`
service which it calls by passing the external id of the product for
which the payment is to be created.

Should anything go wrong, the service throws custom exceptions which
are mapped to handle the erroneous conditions appropriately. In
particular the following two exceptions are handled as follows:

   - `PaymentCreatorNotFoundException` is thrown if no product can be
     found for the give external product id. This condition is mapped
     to status [404] NOT FOUND.
   - `PaymentCreatorDownstreamException` is thrown if a new payment
     can not be created because of a downsteam problem like for
     instance a network problem preventing contacting PublicAPI. This
     condition is mapped to status [500] SYSTEM ERROR.

Any successful new payment is returned in the payload of the response
with status [201] CREATED.