bau: deprecate mockserver
This commit deprecates https://github.com/alphagov/pay-products/pull/166. When
upgrading mockserver, some integration tests were failing. However each test in
that integration test was passing when run in isolation, but some failed when
all the tests were run. Couldn't find a reason as to why that was happening, so
switching to a more reliable mocking library (wiremock).