PP-7067 Alter type of metadata_value column to be VARCHAR(100)
Database migration to alter the type of the metadata_value
column of the products_metadata table from VARCHAR(50) to
VARCHAR(100).

Tested this locally and it did the right thing.

The version of PostgreSQL we use is recent enough increasing the
size of a VARCHAR column does not rewrite the entire table. It
does require an access exclusive lock on the table but in
production products_metadata only has 90 rows so this will be fine.