PP-7067 Increase maximum metadata value length from 50 to 100
Update ProductsMetadataRequestValidator (and its unit tests) to
rely on constants in ExternalMetadata (which is imported from Pay
Java commons) for things like the maximum number of characters
allowed for metadata values associated with a product.

This has no functional effect other than increasing the character
limit for a metadata value from 50 to 100. The necessary database
migration to enable this was commit
32ae783331377b719910ab91df1b521a0f12d59f.

Public API and connector have already been updated to allow 100
characters rather than 50.