PP-7293: Add an endpoint for regenerating an API key for a product
Adds a new endpoint `v1/api/products/{productId}/regenerate-api-key` to the products resource. It is used to generate a new API token and replaces an old API key with the new key for a product with a specified external id. This endpoint is needed for rotating API key for a product because the key can be compromised. It adds two configs items (`emailAddress` and `publicAuthUrl`). Email address is required for identifying who makes a request for generating a new API key. Public Auth URL address is needed to allow Products application to send a request to Public Authentication application for regenerating a new API token. It adds integration tests to cover the following scenarios.

    * API key for payment link successfully replaced
    * API key for prototype link successfully replaced
    * API key for demo payment successfully replaced
    * New API key cannot be generated
    * Product doesn't exist
    * API Key is null
    * API Key is empty