PP-7293: Add connectTimeout and readTimeout to Jersey client
Currently, Jersey client has both connectTimeout and readTimeout set to 0 by default. 0 means infinity. When the client sends a request to another application, it waits until it gets a response. If the request is dropped silently by a firewall or it goes missing, the client will wait forever and never log a timeout exception. This change will add both connectTimeout and readTimeout config items to Products application's config file and are set to 50 seconds by default because Nginx's default timeout is 60 seconds and these timeouts need to be less than 60 seconds to allow Products application to send its response before Nginx cuts off. It also updates RestClientFactoryTest to use JUnit5.