PP-7293: Reduce connectTimeout to 25 seconds
This change updates connectTimeout to use 25 seconds as a default value. It is needed because Jersey client takes 100 seconds to timeout after failing to establish a connection and Nginx cuts this connection after 60 seconds. Reducing from 50 seconds to 25 seconds in connectTimeout causes the timeout to drop from 100 seconds to 50 seconds. readTimeout is not affected by this issue.