PP-8415 Detect PAN in reference field
Log when we see suspected PAN in reference field. This will enable us to remove reference from logs whilst still catching the case when users do this. Products does not log the reference field at all, but once we have this in place can remove the logging from publicapi and connector.

I have chosen to implement this in a separate class to make it easily testable. I have put it in the validator package which is a bit loose, but seems best place for it (happy to discuss this).

For the actual criteria for "looking like a PAN" I have copied and improved on what we do in frontend here and here - length between 12 and 19 inclusive, Luhn valid.