PP-8551 Allow PATCH require_captcha
Add a /v2/api/ endpoint that uses the JSON Patch specification to update
a product.

This is required as the current /v1/ endpoint for updating a product
expects all the fields that can be updated to be provided, and does not
have the ability to update a single field. Adding the new parameter to
this endpoint would risk properties of a product being overwritten if
all consumers of this endpoint aren't kept in sync.

The new /v2/ endpoint is expected to be used for all updates in the
future, but for now it will only support updating the `require_captcha`
property.